Move helpers to a separated module
